renamed res/req variables, changed uint from add to double, and refactord add func, removed t&c in journalcontroller, added 5s timeout

+++ CalculatorService.Core/Models/AddRequest.cs
@@ -8,6 +8,6 @@ namespace CalculatorService.Core.Models
 {
     public class AddRequest
     {
-        public IEnumerable<uint> Addends { get; set; }
+        public IEnumerable<double> Addends { get; set; }
     }
 }
